Handle text management in boxes and separate module
